+++ packages/foundry/contracts/libraries/RoundLib.sol
@@ -24,57 +24,58 @@ library RoundLib {
     // --- Structs ---
 
     struct RoundConfig {
-        uint256 epochDuration; // Duration of each voting epoch (default: 20 minutes)
-        uint256 maxDuration; // Max time before round expires (default: 7 days)
-        uint256 minVoters; // Minimum revealed votes to trigger settlement (default: 3)
-        uint256 maxVoters; // Gas safety cap (default: 1000)
+        uint32 epochDuration; // Duration of each voting epoch (default: 20 minutes)
+        uint32 maxDuration; // Max time before round expires (default: 7 days)
+        uint16 minVoters; // Minimum revealed votes to trigger settlement (default: 3)
+        uint16 maxVoters; // Gas safety cap (default: 1000)
     }
 
     struct Round {
-        uint256 startTime; // When first vote was committed
+        uint48 startTime; // When first vote was committed
         RoundState state;
-        uint256 voteCount; // Total commits across all epochs
-        uint256 revealedCount; // Total revealed votes
-        uint256 totalStake; // Total staked across all voters
-        uint256 upPool; // Total raw stake by UP voters (updated as votes are revealed)
-        uint256 downPool; // Total raw stake by DOWN voters (updated as votes are revealed)
-        uint256 upCount; // Number of UP voters (updated as votes are revealed)
-        uint256 downCount; // Number of DOWN voters (updated as votes are revealed)
+        uint16 voteCount; // Total commits across all epochs
+        uint16 revealedCount; // Total revealed votes
+        uint64 totalStake; // Total staked across all voters
+        uint64 upPool; // Total raw stake by UP voters (updated as votes are revealed)
+        uint64 downPool; // Total raw stake by DOWN voters (updated as votes are revealed)
+        uint16 upCount; // Number of UP voters (updated as votes are revealed)
+        uint16 downCount; // Number of DOWN voters (updated as votes are revealed)
         bool upWins; // Set after settlement
-        uint256 settledAt; // Timestamp when round was settled/tied (for forfeit cutoff)
-        uint256 thresholdReachedAt; // When revealedCount first reached minVoters (0 = not yet)
-        uint256 weightedUpPool; // Epoch-weighted effective stake for UP side (100% epoch-1, 25% epoch-2+)
-        uint256 weightedDownPool; // Epoch-weighted effective stake for DOWN side
+        uint48 settledAt; // Timestamp when round was settled/tied (for forfeit cutoff)
+        uint48 thresholdReachedAt; // When revealedCount first reached minVoters (0 = not yet)
+        uint64 weightedUpPool; // Epoch-weighted effective stake for UP side (100% epoch-1, 25% epoch-2+)
+        uint64 weightedDownPool; // Epoch-weighted effective stake for DOWN side
     }
 
     struct Commit {
         address voter;
-        uint256 stakeAmount;
+        uint64 stakeAmount;
         bytes ciphertext; // tlock-encrypted payload (decryptable after epoch end via drand)
         address frontend; // Frontend operator address (for fee distribution)
-        uint256 revealableAfter; // Epoch end timestamp — reveals allowed after this time
+        uint48 revealableAfter; // Epoch end timestamp — reveals allowed after this time
         bool revealed;
         bool isUp; // Set after reveal
-        uint32 epochIndex; // 0 = epoch 1 (blind, 100% weight), 1 = epoch 2+ (saw results, 25% weight)
+        uint8 epochIndex; // 0 = epoch 1 (blind, 100% weight), 1 = epoch 2+ (saw results, 25% weight)
     }
 
     // --- Epoch weight ---
 
     /// @notice Return epoch weight in BPS: epoch-1 = 10000 (100%), epoch-2+ = 2500 (25%).
-    function epochWeightBps(uint32 epochIndex) internal pure returns (uint256) {
+    function epochWeightBps(uint8 epochIndex) internal pure returns (uint256) {
         return epochIndex == 0 ? 10000 : 2500;
     }
 
     /// @notice Compute epoch-weighted effective stake for a commit.
     function effectiveStake(Commit storage commit) internal view returns (uint256) {
-        return (commit.stakeAmount * epochWeightBps(commit.epochIndex)) / 10000;
+        return (uint256(commit.stakeAmount) * epochWeightBps(commit.epochIndex)) / 10000;
     }
 
     // --- State checks ---
 
     /// @notice Check if a round has expired without reaching settlement.
     function isExpired(Round storage round, uint256 maxDuration) internal view returns (bool) {
-        return round.state == RoundState.Open && round.startTime > 0 && block.timestamp >= round.startTime + maxDuration;
+        return round.state == RoundState.Open && round.startTime > 0
+            && block.timestamp >= uint256(round.startTime) + maxDuration;
     }
 
     /// @notice Check if a round is in a terminal state.
@@ -98,9 +99,9 @@ library RoundLib {
         view
         returns (uint256)
     {
-        uint256 elapsed = commitTimestamp - round.startTime;
+        uint256 elapsed = commitTimestamp - uint256(round.startTime);
         uint256 epochIdx = elapsed / epochDuration;
-        return round.startTime + (epochIdx + 1) * epochDuration;
+        return uint256(round.startTime) + (epochIdx + 1) * epochDuration;
     }
 
     /// @notice Compute the epoch index for a vote committed at the given timestamp (capped at 1).
@@ -111,9 +112,9 @@ library RoundLib {
     function computeEpochIndex(Round storage round, uint256 epochDuration, uint256 commitTimestamp)
         internal
         view
-        returns (uint32)
+        returns (uint8)
     {
-        uint256 elapsed = commitTimestamp - round.startTime;
+        uint256 elapsed = commitTimestamp - uint256(round.startTime);
         uint256 idx = elapsed / epochDuration;
         return idx == 0 ? 0 : 1; // binary two-tier
     }
